requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
openai==1.3.0
transformers==4.36.2
torch==2.1.2
python-dotenv==1.0.0
psutil==7.0.0
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
asyncio-mqtt==0.16.1
